+++ src/main/(shared)/categories/category.service.ts
@@ -2,7 +2,6 @@ import { Injectable } from "@nestjs/common";
 import { slugify } from "@project/utils";
 import { CategoryRepository } from "./category.repository";
 import { CreateCategoryDto } from "./dto/category.dto";
-import { CategoryQueryDto } from "./dto/category.query.dto";
 
 @Injectable()
 export class CategoryService {
@@ -15,7 +14,7 @@ export class CategoryService {
 
     return await this.repository.store({ ...input, slug });
   }
-  async index(options?: CategoryQueryDto) {
-    return await this.repository.findAll(options);
+  async index() {
+    return await this.repository.findAll();
   }
 }