+++ changelog.md
@@ -14,3 +14,5 @@
 
 - (Post Like) handled post like to post (non-tested)
 - (Post Comment) handle indevisual post comments with nested (non-tested)
+- (Community module)
+- (NGO module)